+++ delmo/cli_options.go
@@ -25,7 +25,7 @@ func ParseOptions(args []string) CLIOptions {
 
 OPTIONS:
   -f                    path to the spec file (default: "delmo.yml").
-  -m                    docker-machine to run the tests on (default: "default").
+  -m                    docker-machine to run the tests on.
   --only-build-task     only build the task_image. All other images must be available via docker pull.
   --skip-pull           don't pull the images before building.
   --localhost           IP that will be set to DOCKER_HOST_IP environment variable when not running in a docker-machine.
@@ -36,7 +36,7 @@ OPTIONS:
 	var options CLIOptions
 	options.Usage = usage
 	flags.StringVar(&(options.DelmoFile), "f", "delmo.yml", "")
-	flags.StringVar(&(options.DockerMachine), "m", "default", "")
+	flags.StringVar(&(options.DockerMachine), "m", "", "")
 	flags.BoolVar(&(options.OnlyBuildTask), "only-build-task", false, "")
 	flags.BoolVar(&(options.ParallelExecution), "parallel", false, "")
 	flags.BoolVar(&(options.SkipPull), "skip-pull", false, "")