+++ ci/release_notes.md
@@ -1,3 +1,11 @@
+# Changes
+
+- delmo doesn't run on `docker-machine` by default anymore.
+
 # Features
 
 - `--skip-pull` option to prevent delmo from pulling the images prior to building them. Usefull when network connectivity is bad.
+
+# Bugs
+
+- Correct output when using `--parallel`